UI: Rename decorators file and added coercion to module map
