+++ ui-ngx/src/app/modules/home/pages/notification/rule/escalations.component.ts
@@ -32,7 +32,7 @@ import { AppState } from '@app/core/core.state';
 import { Subject } from 'rxjs';
 import { NonConfirmedNotificationEscalation } from '@shared/models/notification.models';
 import { takeUntil } from 'rxjs/operators';
-import { coerceBoolean } from '@shared/decorators/coerce-boolean';
+import { coerceBoolean } from '@shared/decorators/coercion';
 
 @Component({
   selector: 'tb-escalations-component',
