+++ ui-ngx/src/app/shared/components/notification/template-autocomplete.component.ts
@@ -40,7 +40,7 @@ import {
 } from '@home/pages/notification/template/template-notification-dialog.component';
 import { MatDialog } from '@angular/material/dialog';
 import { MatButton } from '@angular/material/button';
-import { coerceBoolean } from '@shared/decorators/coerce-boolean';
+import { coerceBoolean } from '@shared/decorators/coercion';
 
 @Component({
   selector: 'tb-template-autocomplete',
