+++ ui-ngx/src/app/shared/decorators/public-api.ts
@@ -14,8 +14,6 @@
 /// limitations under the License.
 ///
 
-export * from './components/public-api';
-export * from './decorators/public-api';
-export * from './models/public-api';
-export * from './pipe/public-api';
-export * from './shared.module';
+export * from './coercion';
+export * from './enumerable';
+export * from './tb-inject';
